Edit some CSS stylings including modals and renamed some page routes
