+++ src/components/modals/index.ts
@@ -1,1 +1,2 @@
 export * from './LoginModal'
+export * from './GuideModal'
