+++ src/pages/invites/[id].tsx
@@ -132,7 +132,7 @@ export default function Invite() {
               <h3 className="font-bold text-base text-center">어떤 계정으로 로그인 할까요?</h3>
               <button
                 onClick={() =>
-                  router.push({ pathname: '/auth/login', query: { redirect: `/invite/${router.query.id}` } })
+                  router.push({ pathname: '/auth/login', query: { redirect: `/invites/${router.query.id}` } })
                 }
                 className="block mx-auto btn w-full max-w-xs mt-4 bg-primary text-white"
               >
